requirements.txt
# LLM & LangChain
langchain>=0.3.25
langchain-core>=0.3.62
langchain-text-splitters>=0.3.8
langchain-ollama>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.24

# Vector DB
chromadb>=1.0.9

# Embeddings
sentence-transformers>=4.1.0

# Web Search
langchain-tavily>=0.1.0
tavily-python>=0.5.0

# Document Processing
pypdf>=5.5.0
PyMuPDF>=1.25.0
python-docx>=1.1.0

# UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
